Set publisher and placeOfPublication in ReShare request
I have determined experimentally that these are included as `rft.pub`
and `rft.place` in the v1.0 OpenURLs generated by the New School's
discovery layer.

I can't tell if this is standard, because -- and I still can't quite
believe this -- the ANSI/NISO Z39.88 document doesn't include a list
of fields.