Changes to HTML form
* Email address removed (and no longer mandatory)
* New date-needed-by field.

As required for PR-502.

The latter requires new support for the back-end field `neededBy`. But
this is defines as being of type LocalDate, which means there are
certain constraints on what values can be accepted: it seems
experimentally that ISO 8601 format is used. We may end up needing to
do undesirably clever things to enforce this.

Corresponding changes to documentation and tests.