Fix two of three test errors
These were caused by atempting to fetch a list of candidate pickup
locations when no `svc.pickupLocation` was specified. If have now
added support for an undocumented `svc.noPickupLocation` key which
tests can use to tell the resolver not to fetch such a list, but to
proceed with an empty list; and this key is included in the two tests
that were failing for that reason.

The remaining test seems to be for a different reason.

Fixes part of PR-641.