Fix problems with HTML of pickup-location form
This was requesting two JavaScript files that it didn't need and which
we don't have; and also requesting an image from the wrong
location. As a result each real request was causing three more dumb
ones, doomed to fail -- and each was causing a re-request of the
backend service's list of pickup locations.

We now make only 40% as many backend requests as before: one request
for pickup locations instead of four, plus the actual post once that
is resolved.

As a bonus, we now included the scroll-arrow image in the dropdown,
which we were previously not rendering