Update linting to work with ESM
The standardx eslint wrapper previously used has not been updated for a
couple years and brought in an earlier version of eslint unaware of
NodeJS support for modules.

Luckily the eslint config it was based on is directly importable. I also
brought in one for mocha to better lint our tests.

For now I disabled a couple of rules regarding setup directly in
describe() rather than wrapped in before() and where one file is
nesting tests. They do offer some rationale and we should probably
consider addressing these, especially as the latter might be ignoring
some tests:

https://github.com/lo1tuma/eslint-plugin-mocha/blob/master/docs/rules/no-setup-in-describe.md
https://github.com/lo1tuma/eslint-plugin-mocha/blob/master/docs/rules/no-nested-tests.md