Avoid internal server error
OpenURLs with no `ntries` defined could cause this as the default
value assigned in code was the number 0 (not the string '0'), and the
`replaceAll` method is not defined on numbers.

While I was there I fixed the code that increments this number on
subsequent tries at submitting the form. It was Working By Accident
before, as the string '0' is not considered greater than the number 0
but '01' is.

Fixes PR-1646.