Set `json` to empty object if JSON parsing fails
This allows the rest of the code to run after POSTing the request to
ReShare fails. At present this is impossible, as a hard guard returns
from the function early, so that the `bad.html` template is never
used, but if we remove that guard (why was it added?) all will be
well.