Revert "Define inline partial template "field", use for volume"
This reverts commit f43d1d56f18ad127062948f744a92811629e5bef.

Turns out that this isn't going to work, as Handlebars is too limited:
there is no way to get the current values of a field into the "field"
partial:
* You can't pass the value in directly, because invocations that nest
one {{}} block inside another are syntactically invalid.
* You can't have the partial generate the value by looking up
specified name, because there is no syntax (at least none that's
documented) for indirecting through the value of a parameter.

So we'll just have to go back to writing out each field definition by
hand.