Hide/requires fields irrelevant/relevant to the chosen service type (#49)
Fixes PR-1657 and PR-1659.

These two closely related issues both pertain to how things are different when the user requests a loan (the old operation) and a copy. As these changes, the set of fields that are mandatory changes, with those changes shown dynamically in the HTML form and enforced in the resolver. Irrelevant fields are disabled and greyed out.