Blank request form correctly displays Default copyright type.
Fixes PR-1716.

The problem here is mostly that the copyright type stored as the
default has an ID that does not match any of the actual copyright
types. We work around this by looking up the _value_ of the default
copyright type in the list of candidate types, and using its ID.