First stab at dynamically setting "PLease provide a ..." messages
Towards PR-1717.

Currently decides whether a message is needed based on the data that
was used to populate the form. Should instead use the present state of
the form, so the messages appear and disappear as you edit. This will
also make the code simpler, as it will no longer be necessary to pass
in the stringified data.